# [Agent] Optimize epc_id queries